requirements/requirements.txt
click==7.1.2
pycryptodome==3.9.9